+++ components/mediation-initializer/org.wso2.carbon.mediation.initializer/pom.xml
@@ -115,7 +115,6 @@
                         </CAPP_MANAGER-RequiredServices>
                         <Export-Package>
                             org.wso2.carbon.mediation.initializer.*,
-                            org.wso2.carbon.mediation.initializer.request.processors
                         </Export-Package>
                         <Import-Package>
                             *;resolution:=optional
